adio-api-main/requirements.txt
alembic==1.10.2
Flask==2.2.2
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
Flask_Migrate==4.0.4
Flask_SQLAlchemy==2.5.1
PyJWT==2.6.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
psycopg2-binary
flask-socketio
websockets